refactoring and fancy readme
