+++ package.json
@@ -14,6 +14,7 @@
   "author": "nodejs",
   "license": "ISC",
   "dependencies": {
+    "@types/node": "^22.9.1",
     "dotenv": "^16.4.5",
     "express": "^4.17.1"
   },
