+++ src/lib/helpers.ts
@@ -1,3 +1,16 @@
+export const makeQueryFromObject = (obj: {
+  [x: string]: any;
+  start_date?: string;
+  end_date?: string;
+  workspace_id?: string;
+}) => {
+  return Object.keys(obj)
+    .map((key, idx) => {
+      return `${idx === 0 ? "?" : ""}${key}=${obj[key]}`;
+    })
+    .join("&");
+};
+
 export async function fetchJSON(
   url: string,
   options: RequestInit = {}
