+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "target": "ES2020",
     "module": "CommonJS",
-    "lib": ["ES2020"],
+    "lib": ["dom", "ES2020"],
     "outDir": "dist",
     "strict": true,
     "esModuleInterop": true,
